geotiff tests: consolidate coord cluster (#2428)
Fold seven top-level files into read/test_coords.py:

- test_coord_regularity_1720.py
- test_coords_1813.py
- test_coords_to_transform_3d_1643.py
- test_int_coord_sentinel_2087.py
- test_int_coords_round_trip_hotfix_1962.py
- test_multi_tiepoint_validation_2117.py
- test_rational_zero_denominator_2313.py

The boundary file test_non_uniform_coords_alias_2215.py also folds
cleanly because the alias resolution layers onto the same writer-side
_coords_to_transform validation already covered in section 5; it lands
in the same section.

Sections in the extended file:

1. coords_from_pixel_geometry / coords_from_geo_info helpers.
2. _extract_transform multi-tiepoint consistency.
3. Zero-denominator RATIONAL / SRATIONAL rejection.
4. Descending / ascending coord round trip (pre-existing from #2390).
5. _coords_to_transform writer-side validation: regularity, 3D layouts,
   and alias-aware NonUniformCoordsError.
6. Integer-coord round trip and the _NO_GEOREF_KEY marker. Integration
   sub-cluster gated on XRSPATIAL_RUN_INTEGRATION=1.

CLUSTER_AUDIT_COORDS.md maps every old file::test to its new
file::test_id. The audit file is deleted on a final commit before merge.

Cluster 4 of long-tail epic #2424. Tests-only -- no source changes.
